Update usb-redirection.md
The first part of the documentation is a bit unclear as to which icon is refered to for example here:
2. On the client interface, click the icon to list the locally attached USB devices. 

I suggest to use the same ICON and text as this sentence:
After you log into your WorkSpace, select the Devices icon Devices icon , and reinstall the USB driver. Alternatively, you can invoke the following PowerShell script (using administrator privileges). 


See my edits